#7091: add endpoint to get list of layers: integration with kcapi, insights-api

+++ src/main/java/io/kontur/disasterninja/dto/layer/LayerDetailsDto.java
@@ -7,8 +7,8 @@ import lombok.Data;
 public class LayerDetailsDto {
 
     private final String id;
-    private final Double maxZoom;
-    private final Double minZoom;
+    private final Integer maxZoom;
+    private final Integer minZoom;
     private final LayerSourceDto source;
 
     public static LayerDetailsDto fromLayer(Layer layer) {
